favorites exception implemented

+++ src/main/java/service/FavoriteService.java
@@ -1,6 +1,8 @@
 package service;
 
 import dao.FavoriteDAO;
+import exception.AlreadyMarkedFavoriteException;
+import exception.FavoriteException;
 import model.Product;
 
 import java.util.List;
@@ -16,17 +18,27 @@ public class FavoriteService {
     {
         if(favoriteDAO.isFavorite(userId,productId))
         {
-            System.out.println("Already in favorites");
+           throw new AlreadyMarkedFavoriteException("Product is already marked as favorites");
         }
         favoriteDAO.addFavorite(userId,productId);
     }
 
     public  void removeFromFavorites(int userId,int productId)
     {
-        favoriteDAO.removeFavorite(userId,productId);
+        if (!favoriteDAO.isFavorite(userId, productId)) {
+            throw new FavoriteException("Product not in favorites");
+        }
+
+        favoriteDAO.removeFavorite(userId, productId);
     }
+
     public List<Product> viewFavorites(int userId) {
-        return favoriteDAO.getFavByUserId(userId);
+        List<Product> favorites= favoriteDAO.getFavByUserId(userId);
+        if(favorites.isEmpty() || favorites==null)
+        {
+            throw new FavoriteException("Favorite Product Not found");
+        }
+        return favorites;
     }
 
 }
